Issue #12 Fix crash on duplicate block creation

+++ modules/blockcreator/blockmanager.go
@@ -23,6 +23,10 @@ func (m *BlockCreator) submitBlock(blockToSubmit types.Block) error {
 		m.log.Println("Created an unsolved block - submission appears to be incorrect")
 		return err
 	}
+	if err == modules.ErrBlockKnown {
+		m.log.Println("Created a block that is already known in the consensusset")
+		return err
+	}
 	if err != nil {
 		m.tpool.PurgeTransactionPool()
 		m.log.Critical("ERROR: an invalid block was submitted:", err)
